Add hasAnInternshipAgreement to the companies table
