+++ src/pages/Admin/Companies/Columns/constants.ts
@@ -4,6 +4,7 @@ export enum COMPANY_LIST_COLUMNS_VALUES {
   BUSINESS_SECTOR = "businessSector",
   CUIT = "cuit",
   UPDATED_AT = "updatedAt",
+  HAS_AN_INTERNSHIP_AGREEMENT = "hasAnInternshipAgreement",
   STATE = "state",
   ACTIONS = "actions"
 }
@@ -14,6 +15,7 @@ export const COMPANY_LIST_COLUMNS = [
   COMPANY_LIST_COLUMNS_VALUES.BUSINESS_SECTOR,
   COMPANY_LIST_COLUMNS_VALUES.CUIT,
   COMPANY_LIST_COLUMNS_VALUES.UPDATED_AT,
+  COMPANY_LIST_COLUMNS_VALUES.HAS_AN_INTERNSHIP_AGREEMENT,
   COMPANY_LIST_COLUMNS_VALUES.STATE,
   COMPANY_LIST_COLUMNS_VALUES.ACTIONS
 ];
