+++ src/pages/Admin/Companies/ListHeader/Item/styles.module.scss
@@ -23,6 +23,10 @@
     grid-area: updatedAt;
   }
 
+  &.hasAnInternshipAgreement {
+    grid-area: hasAnInternshipAgreement;
+  }
+
   &.state {
     grid-area: state;
   }
